+++ packages/autocomplete-js/Autocomplete.js
@@ -2,6 +2,7 @@ import AutocompleteCore from '../autocomplete/AutocompleteCore.js'
 import uniqueId from '../autocomplete/util/uniqueId.js'
 import getRelativePosition from '../autocomplete/util/getRelativePosition.js'
 import debounce from '../autocomplete/util/debounce.js'
+import getAriaLabel from '../autocomplete/util/getAriaLabel.js'
 
 // Creates a props object with overridden toString function. toString returns an attributes
 // string in the format: `key1="value1" key2="value2"` for easy use in an HTML string.
@@ -42,6 +43,8 @@ class Autocomplete {
       getResultValue = result => result,
       renderResult,
       debounceTime = 0,
+      resultListLabel,
+      submitOnEnter = false,
     } = {}
   ) {
     this.root = typeof root === 'string' ? document.querySelector(root) : root
@@ -54,6 +57,8 @@ class Autocomplete {
     if (typeof renderResult === 'function') {
       this.renderResult = renderResult
     }
+    this.resultListLabel = resultListLabel
+    this.submitOnEnter = submitOnEnter
 
     const core = new AutocompleteCore({
       search,
@@ -67,6 +72,7 @@ class Autocomplete {
       onHide: this.handleHide,
       onLoading: this.handleLoading,
       onLoaded: this.handleLoaded,
+      submitOnEnter: this.submitOnEnter,
     })
     if (debounceTime > 0) {
       core.handleInput = debounce(core.handleInput, debounceTime)
@@ -92,6 +98,14 @@ class Autocomplete {
     this.input.setAttribute('aria-expanded', 'false')
 
     this.resultList.setAttribute('role', 'listbox')
+
+    const resultListAriaLabel = getAriaLabel(this.resultListLabel)
+    resultListAriaLabel &&
+      this.resultList.setAttribute(
+        resultListAriaLabel.attribute,
+        resultListAriaLabel.content
+      )
+
     this.resultList.style.position = 'absolute'
     this.resultList.style.zIndex = '1'
     this.resultList.style.width = '100%'