+++ packages/autocomplete-vue/Autocomplete.vue
@@ -42,6 +42,7 @@ import AutocompleteCore from '../autocomplete/AutocompleteCore.js'
 import uniqueId from '../autocomplete/util/uniqueId.js'
 import getRelativePosition from '../autocomplete/util/getRelativePosition.js'
 import debounce from '../autocomplete/util/debounce.js'
+import getAriaLabel from '../autocomplete/util/getAriaLabel'
 
 export default {
   name: 'Autocomplete',
@@ -72,6 +73,14 @@ export default {
       type: Number,
       default: 0,
     },
+    resultListLabel: {
+      type: String,
+      default: undefined,
+    },
+    submitOnEnter: {
+      type: Boolean,
+      default: false,
+    },
   },
 
   data() {
@@ -85,6 +94,7 @@ export default {
       onHide: this.handleHide,
       onLoading: this.handleLoading,
       onLoaded: this.handleLoaded,
+      submitOnEnter: this.submitOnEnter,
     })
     if (this.debounceTime > 0) {
       core.handleInput = debounce(core.handleInput, this.debounceTime)
@@ -143,10 +153,13 @@ export default {
     },
     resultListProps() {
       const yPosition = this.position === 'below' ? 'top' : 'bottom'
+      const ariaLabel = getAriaLabel(this.resultListLabel)
+
       return {
         id: this.resultListId,
         class: `${this.baseClass}-result-list`,
         role: 'listbox',
+        [ariaLabel?.attribute]: ariaLabel?.content,
         style: {
           position: 'absolute',
           zIndex: 1,