+++ CHANGELOG.md
@@ -1,7 +1,25 @@
+# Changelog
+
+## v2.4.0 (January 9, 2023)
+
+### autocomplete-core
+
+- Added `submitOnEnter` option to control whether or not results should be submitted immediately after selecting them by pressing <kbd>Enter</kbd> (dpxgit, [@dpxgit](https://github.com/dpxgit))
+
+### autocomplete-js
+
+- Added `resultListLabel` option to provide `aria-label` or `aria-labelledby` attribute for result list (dpxgit, [@dpxgit](https://github.com/dpxgit))
+
+### autocomplete-vue
+
+- Added `resultListLabel` prop to provide `aria-label` or `aria-labelledby` attribute for result list (dpxgit, [@dpxgit](https://github.com/dpxgit))
+
 ## v2.3.0 (November 2, 2022)
+
 ### autocomplete-js
 
 - Fixed W3C validation error regarding the attribute `autocorrect` by changing it to be optional (dpxgit, [@dpxgit](https://github.com/dpxgit))
+
 ## v2.2.0 (May 28, 2020)
 
 ### autocomplete-vue